New finger postion change tracking tracking method
Using distance from a critical point on the finger has a fundamental problem. When the camera angle change or the distance changes from your hand to the camera, the distance will be different. It will create problems in the game. In this update, I will take three critical points from each finger, then use that to calculate the angle of the finger position.